+++ js/data/api.js
@@ -1,0 +1,1 @@
+// Получение и отправка данных
