+++ js/form/file-upload.js
@@ -1,0 +1,1 @@
+// Загрузка файлов
