+++ js/form/form.js
@@ -1,0 +1,1 @@
+// Заполнение информации
