+++ js/map/filter.js
@@ -1,0 +1,1 @@
+// Фильтры карты
