+++ js/map/map.js
@@ -1,0 +1,1 @@
+// Основной модуль для работы с картой
