+++ js/utils/app-state.js
@@ -1,0 +1,1 @@
+// 1. Переключение неактивного/активного состояния страницы при загрузке карты.
