+++ js/utils/validation.js
@@ -1,0 +1,1 @@
+// Валидации всех полей формы
